user_order_api/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==4.2.23
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
flake8==7.3.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
oauthlib==3.3.1
psycopg==3.2.9
psycopg-binary==3.2.9
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
